+++ pandas/tests/base/test_value_counts.py
@@ -30,7 +30,7 @@ def test_value_counts(index_or_series_obj):
     result = obj.value_counts()
 
     counter = collections.Counter(obj)
-    expected = pd.Series(dict(counter.most_common()), dtype=np.int64, name=obj.name)
+    expected = pd.Series(dict(counter.most_common()), dtype="Int64", name=obj.name)
     expected.index = expected.index.astype(obj.dtype)
     if isinstance(obj, pd.MultiIndex):
         expected.index = pd.Index(expected.index)
@@ -67,7 +67,7 @@ def test_value_counts_null(null_obj, index_or_series_obj):
     # because np.nan == np.nan is False, but None == None is True
     # np.nan would be duplicated, whereas None wouldn't
     counter = collections.Counter(obj.dropna())
-    expected = pd.Series(dict(counter.most_common()), dtype=np.int64)
+    expected = pd.Series(dict(counter.most_common()), dtype="Int64")
     expected.index = expected.index.astype(obj.dtype)
 
     result = obj.value_counts()
@@ -80,7 +80,7 @@ def test_value_counts_null(null_obj, index_or_series_obj):
 
     # can't use expected[null_obj] = 3 as
     # IntervalIndex doesn't allow assignment
-    new_entry = pd.Series({np.nan: 3}, dtype=np.int64)
+    new_entry = pd.Series({np.nan: 3}, dtype="Int64")
     expected = expected.append(new_entry)
 
     result = obj.value_counts(dropna=False)
@@ -96,7 +96,7 @@ def test_value_counts_inferred(index_or_series):
     klass = index_or_series
     s_values = ["a", "b", "b", "b", "b", "c", "d", "d", "a", "a"]
     s = klass(s_values)
-    expected = Series([4, 3, 2, 1], index=["b", "a", "d", "c"])
+    expected = Series([4, 3, 2, 1], index=["b", "a", "d", "c"], dtype="Int64")
     tm.assert_series_equal(s.value_counts(), expected)
 
     if isinstance(s, Index):
@@ -110,17 +110,17 @@ def test_value_counts_inferred(index_or_series):
     # don't sort, have to sort after the fact as not sorting is
     # platform-dep
     hist = s.value_counts(sort=False).sort_values()
-    expected = Series([3, 1, 4, 2], index=list("acbd")).sort_values()
+    expected = Series([3, 1, 4, 2], index=list("acbd"), dtype="Int64").sort_values()
     tm.assert_series_equal(hist, expected)
 
     # sort ascending
     hist = s.value_counts(ascending=True)
-    expected = Series([1, 2, 3, 4], index=list("cdab"))
+    expected = Series([1, 2, 3, 4], index=list("cdab"), dtype="Int64")
     tm.assert_series_equal(hist, expected)
 
     # relative histogram.
     hist = s.value_counts(normalize=True)
-    expected = Series([0.4, 0.3, 0.2, 0.1], index=["b", "a", "d", "c"])
+    expected = Series([0.4, 0.3, 0.2, 0.1], index=["b", "a", "d", "c"], dtype="float64")
     tm.assert_series_equal(hist, expected)
 
 
@@ -136,16 +136,16 @@ def test_value_counts_bins(index_or_series):
 
     s1 = Series([1, 1, 2, 3])
     res1 = s1.value_counts(bins=1)
-    exp1 = Series({Interval(0.997, 3.0): 4})
+    exp1 = Series({Interval(0.997, 3.0): 4}, dtype="Int64")
     tm.assert_series_equal(res1, exp1)
     res1n = s1.value_counts(bins=1, normalize=True)
-    exp1n = Series({Interval(0.997, 3.0): 1.0})
+    exp1n = Series({Interval(0.997, 3.0): 1.0}, dtype="float64")
     tm.assert_series_equal(res1n, exp1n)
 
     if isinstance(s1, Index):
         tm.assert_index_equal(s1.unique(), Index([1, 2, 3]))
     else:
-        exp = np.array([1, 2, 3], dtype=np.int64)
+        exp = np.array([1, 2, 3], dtype="Int64")
         tm.assert_numpy_array_equal(s1.unique(), exp)
 
     assert s1.nunique() == 3
@@ -153,22 +153,24 @@ def test_value_counts_bins(index_or_series):
     # these return the same
     res4 = s1.value_counts(bins=4, dropna=True)
     intervals = IntervalIndex.from_breaks([0.997, 1.5, 2.0, 2.5, 3.0])
-    exp4 = Series([2, 1, 1, 0], index=intervals.take([0, 3, 1, 2]))
+    exp4 = Series([2, 1, 1, 0], index=intervals.take([0, 3, 1, 2]), dtype="Int64")
     tm.assert_series_equal(res4, exp4)
 
     res4 = s1.value_counts(bins=4, dropna=False)
     intervals = IntervalIndex.from_breaks([0.997, 1.5, 2.0, 2.5, 3.0])
-    exp4 = Series([2, 1, 1, 0], index=intervals.take([0, 3, 1, 2]))
+    exp4 = Series([2, 1, 1, 0], index=intervals.take([0, 3, 1, 2]), dtype="Int64")
     tm.assert_series_equal(res4, exp4)
 
     res4n = s1.value_counts(bins=4, normalize=True)
-    exp4n = Series([0.5, 0.25, 0.25, 0], index=intervals.take([0, 3, 1, 2]))
+    exp4n = Series(
+        [0.5, 0.25, 0.25, 0], index=intervals.take([0, 3, 1, 2]), dtype="float64"
+    )
     tm.assert_series_equal(res4n, exp4n)
 
     # handle NA's properly
     s_values = ["a", "b", "b", "b", np.nan, np.nan, "d", "d", "a", "a", "b"]
     s = klass(s_values)
-    expected = Series([4, 3, 2], index=["b", "a", "d"])
+    expected = Series([4, 3, 2], index=["b", "a", "d"], dtype="Int64")
     tm.assert_series_equal(s.value_counts(), expected)
 
     if isinstance(s, Index):
@@ -180,7 +182,7 @@ def test_value_counts_bins(index_or_series):
     assert s.nunique() == 3
 
     s = klass({}) if klass is dict else klass({}, dtype=object)
-    expected = Series([], dtype=np.int64)
+    expected = Series([], dtype="Int64")
     tm.assert_series_equal(s.value_counts(), expected, check_index_type=False)
     # returned dtype differs depending on original
     if isinstance(s, Index):
@@ -216,7 +218,7 @@ def test_value_counts_datetime64(index_or_series):
     idx = pd.to_datetime(
         ["2010-01-01 00:00:00", "2008-09-09 00:00:00", "2009-01-01 00:00:00"]
     )
-    expected_s = Series([3, 2, 1], index=idx)
+    expected_s = Series([3, 2, 1], index=idx, dtype="Int64")
     tm.assert_series_equal(s.value_counts(), expected_s)
 
     expected = np_array_datetime64_compat(
@@ -240,7 +242,7 @@ def test_value_counts_datetime64(index_or_series):
 
     result = s.value_counts(dropna=False)
     expected_s[pd.NaT] = 1
-    tm.assert_series_equal(result, expected_s)
+    tm.assert_series_equal(result, expected_s.astype("Int64"))
 
     unique = s.unique()
     assert unique.dtype == "datetime64[ns]"
@@ -261,7 +263,7 @@ def test_value_counts_datetime64(index_or_series):
     td = klass(td, name="dt")
 
     result = td.value_counts()
-    expected_s = Series([6], index=[Timedelta("1day")], name="dt")
+    expected_s = Series([6], index=[Timedelta("1day")], name="dt", dtype="Int64")
     tm.assert_series_equal(result, expected_s)
 
     expected = TimedeltaIndex(["1 days"], name="dt")
