+++ pandas/tests/io/pytables/test_store.py
@@ -1990,7 +1990,8 @@ class TestHDFStore:
                     "int64": 1,
                     "object": 1,
                     "datetime64[ns]": 2,
-                }
+                },
+                dtype="Int64",
             )
             result = result.sort_index()
             expected = expected.sort_index()
